Make obj3_t::remove_key() populate _dict if needed
Previously, remove_key() avoided creating _dict if it didn’t already
exist. This made it a noop if the object was not a dictionary, instead
of wiping out array/scalar content.

Unfortunately this also made it a noop for an obj which is a reference
and whose _dict property hadn’t been populated with a pointer to the
referenced object’s _dict yet.

operator() does coerce the obj into a dictionary. Change remove_key() to
have the same behavior.